+++ business/business/ecommerce/ItemRequisicao.cs
@@ -1,6 +1,7 @@
 ﻿using business.business;
 using business.business.Elementos.element;
 using Newtonsoft.Json;
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace business.ecommerce
@@ -8,12 +9,12 @@ namespace business.ecommerce
     public class ItemRequisicao : BaseModel
     {
         public int Quantidade { get; set; }
-        public ulong RequisicaoId { get; set; }
+        public Int64 RequisicaoId { get; set; }
         [Required]
-        public decimal PrecoUnitario { get; private set; }
+        public Int64 PrecoUnitario { get; private set; }
         [JsonIgnore]
         public virtual Requisicao Requisicao { get; set; }
-        public ulong ElementoId { get; set; }
+        public Int64 ElementoId { get; set; }
         [JsonIgnore]
         public virtual Elemento Elemento { get; set; }
